+++ Assets/chocopoi/DressingTools/Editor/Containers.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 5be3d3e213639d343bff6611f703234a
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
