+++ Assets/chocopoi/DressingTools/Editor/Containers/DTDynamicBone.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using UnityEngine;
 
-namespace Chocopoi.DressingTools
+namespace Chocopoi.DressingTools.Containers
 {
     public class DTDynamicBone
     {
