+++ Assets/chocopoi/DressingTools/Editor/Reporting.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 1d1f7b5618b3c8244a46ed780cc42f55
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
