+++ Assets/chocopoi/DressingTools/Editor/Reporting/DressCheckResult.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using UnityEngine;
 
-namespace Chocopoi.DressingTools
+namespace Chocopoi.DressingTools.Reporting
 {
     public enum DressCheckResult
     {
