+++ Assets/chocopoi/DressingTools/Editor/Reporting/DressReport.cs
@@ -4,21 +4,22 @@ using UnityEditor;
 using UnityEditor.Animations;
 using UnityEngine;
 using VRC.SDK3.Dynamics.PhysBone.Components;
+using Chocopoi.DressingTools.Containers;
 
-namespace Chocopoi.DressingTools
+namespace Chocopoi.DressingTools.Reporting
 {
     public class DressReport
     {
-        private static readonly IDressCheckRule[] rules = new IDressCheckRule[]
+        private static readonly Rules.IDressCheckRule[] rules = new Rules.IDressCheckRule[]
         {
-            new NotAPrefabRule(),
-            new NoMissingScriptsRule(),
-            new ExistingPrefixSuffixRule(),
-            new FindAvatarDynamicsRule(),
-            new ArmatureRule(),
-            new GroupRootObjectsRule(),
-            new FindClothesDynamicsRule(),
-            new GroupClothesDynamicsRule()
+            new Rules.NotAPrefabRule(),
+            new Rules.NoMissingScriptsRule(),
+            new Rules.ExistingPrefixSuffixRule(),
+            new Rules.FindAvatarDynamicsRule(),
+            new Rules.ArmatureRule(),
+            new Rules.GroupRootObjectsRule(),
+            new Rules.FindClothesDynamicsRule(),
+            new Rules.GroupClothesDynamicsRule()
         };
 
         private static AnimatorController testModeAnimationController;
@@ -136,7 +137,7 @@ namespace Chocopoi.DressingTools
                 targetClothes = settings.clothesToDress;
             }
 
-            foreach (IDressCheckRule rule in rules)
+            foreach (Rules.IDressCheckRule rule in rules)
             {
                 if (!rule.Evaluate(report, settings, targetAvatar, targetClothes))
                 {
