+++ Assets/chocopoi/DressingTools/Editor/Rules/GroupClothesDynamicsRule.cs
@@ -3,8 +3,10 @@ using System.Collections.Generic;
 using UnityEditor;
 using UnityEngine;
 using VRC.SDK3.Dynamics.PhysBone.Components;
+using Chocopoi.DressingTools.Reporting;
+using Chocopoi.DressingTools.Containers;
 
-namespace Chocopoi.DressingTools
+namespace Chocopoi.DressingTools.Rules
 {
     public class GroupClothesDynamicsRule : IDressCheckRule
     {
