+++ Assets/chocopoi/DressingTools/Editor/Rules/NotAPrefabRule.cs
@@ -2,8 +2,9 @@
 using System.Collections.Generic;
 using UnityEditor;
 using UnityEngine;
+using Chocopoi.DressingTools.Reporting;
 
-namespace Chocopoi.DressingTools
+namespace Chocopoi.DressingTools.Rules
 {
     public class NotAPrefabRule : IDressCheckRule
     {
