+++ Assets/chocopoi/DressingTools/Editor/Translation.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: e556b462818fdb2469c5f74b552694df
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
